backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pandas==2.2.1
openpyxl==3.1.2
python-dateutil==2.8.2 